Start fetching the currencies directly from endpoint
This commit updates the way I load the currencies table into the component.
I want to eventually transition to having the Go method generate the currency
table every now and then. For this, I need first to transition the React
component to start reading from a public json file instead of a hardcoded
type with values.

Eventually the plan is to update the go function to read the file location
from an env variable, and just update it by running it. This means that
the file will need to be removed and added to gitignore once thats done.

Pending to add tests to all of this once I have some time.